[Android] Change the API for ApplicationStatus in base dir.
Chromium class inherits from Application directly. This is not suitable
for Crosswalk embedding API. Instead, define a new API which can pass Application
instance as the parameter instead of asking users' application inheriting from
Chromium Application.

It is caused by this upstream CL: https://codereview.chromium.org/159173002

M49 rebasing notes: adjusted to the following CLs:
 * https://codereview.chromium.org/1649963002
 * https://codereview.chromium.org/1678103006

M50 rebasing notes: adjusted to the following CLs:
 * https://codereview.chromium.org/1605993004
